simplify interface and move data structures
